AI_Projects/AI/Baby_Book_Writer/requirements.txt
groq>=0.4.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
